Refactor the game map to permit exposing it to Lua
This adds map[{x,y}] and map[x][y] as valid ways to access the individual tiles of the map, rather than using map.set_terrain and map.get_terrain. This renders get_terrain redundant but not set_terrain (as the latter offers a third parameter).

Special locations as viewed through the map userdata are now iterable, and the length function also works.